build: Clean up build system and output.
Try to reduce repeated variable names, and use version numbers in the
path.  It would be nice if Ant spoke Git, so we could use information
about the state of the build tree in build artifacts.

Also, tweak the output of `run`.  From a user-experience perspective,
Ant output is scary.  Trying to minimise the volume of that is good.

Signed-off-by: Jashank Jeremy <jashankj@cse.unsw.edu.au>